+++ app/assets/components/editor/style.less
