+++ app/assets/components/global-search/style.less
