+++ app/assets/components/header/index.jsx
@@ -3,7 +3,7 @@ import { Avatar, Icon } from 'antd';
 import { Link } from 'dva/router';
 import GlobalSearch from '../global-search';
 
-import './index.less';
+import './style.less';
 
 const user = window.context.user;
 class Header extends React.Component {
