+++ app/assets/components/info/style.less
