+++ app/assets/components/schema-editor/index.jsx
@@ -3,13 +3,13 @@ import autobind from 'autobind-decorator';
 import set from 'lodash/set';
 import { Modal, Icon, Tooltip } from 'antd';
 import { Controlled as CodeMirror } from 'react-codemirror2';
-import json2schema from '../../utils/json2schemav2';
+import json2schema from '../../utils/json2schema';
 import schema2json from '../../utils/schema2json';
 import JSON5 from 'json5';
 import stringToPath from '../../utils/stringToPath';
 import BulkEditor from '../bulk-editor/index';
 
-import './index.less';
+import './style.less';
 
 // CodeMirror packages
 require('codemirror/mode/javascript/javascript');
