+++ app/assets/components/schema-editor/style.less
