+++ app/assets/pages/api-test-page/components/my-tabs/style.less
