+++ app/assets/pages/collection-api-list-page/style.less
