+++ app/assets/pages/collection-apis-page/style.less
