+++ app/assets/pages/collection-edit-page/style.less
