+++ app/assets/pages/collection-page/style.less
