+++ app/assets/pages/collection-setting-page/index.jsx
@@ -7,7 +7,7 @@ import cloneDeep from 'lodash/cloneDeep';
 import { Link } from 'dva/router';
 import { TypeColorMap } from '../../utils/constants';
 
-import './index.less';
+import './style.less';
 
 const FormItem = Form.Item;
 const { TextArea } = Input;
