+++ app/assets/pages/collection-setting-page/style.less
