+++ app/assets/pages/collections-page/style.less
