+++ app/assets/pages/project-edit-page/index.jsx
@@ -6,7 +6,7 @@ import { connect } from 'dva';
 import { isOwner } from '../../utils/utils';
 import Layout from '../../layout/default.jsx';
 import BulkEditor from '../../components/bulk-editor/index';
-import './index.less';
+import './style.less';
 
 const createForm = Form.create;
 const FormItem = Form.Item;
