+++ app/assets/pages/projects-page/index.jsx
@@ -5,7 +5,7 @@ import { connect } from 'dva';
 import { Link } from 'dva/router';
 import Layout from '../../layout/default.jsx';
 
-import './index.less';
+import './style.less';
 
 class ProjectsPage extends React.PureComponent {
   constructor(props) {
